+++ src/models/index.ts
@@ -1,0 +1,8 @@
+import Team from './teams'
+import Event from './events'
+import Match from './matches'
+
+export type EventData = Event;
+export type MatchData = Match;
+export type TeamData = Team;
+
