+++ src/store/actions.ts
@@ -1,9 +1,10 @@
 import {actions as events} from './events'
 import {actions as matches} from './matches'
 import {actions as user } from './user' 
-
+import {actions as teams} from './teams'
 export {
     events,
     matches,
-    user
+    user,
+    teams
 }
