+++ src/store/matches/types.ts
@@ -1,25 +1,10 @@
 import { Types } from '../../database'
+import {MatchData as Model} from '../../models'
+
+export type MatchData = Model;
 
 export const GET_MATCHES = "GET_MATCHES";
 export const ADD_MATCHES = 'ADD_MATCHES';
-
-export type MatchType = "QUALIFIER" | "R16" | "R8" | "R4" | "R2";
-export interface MatchData extends Types.DocumentData {
-    DOCUMENT_TYPE : typeof Types.Collections.match;
-
-    eventId: string;
-    blueTeamTop: string;
-    blueTeamBottom: string;
-    redTeamTop: string;
-    redTeamBottom: string;
-    blueScore ?: number;
-    redScore ?: number;
-    type : MatchType;
-    time : string;
-    matchNumber : number;
-    completed : boolean;
-    id:number;
-}
 
 export type MatchDoc = Types.DocumentBase<MatchData>;
