+++ src/store/reducers.ts
@@ -2,5 +2,6 @@ import {combineReducers} from 'redux';
 import {eventsReducer as events} from './events/reducers';
 import {matchesReducer as matches} from './matches/reducers';
 import { userReducer as user } from './user/reducers'
+import { teamsReducer as teams} from './teams/reducers';
 
-export const rootReducer = combineReducers({events, matches, user});
+export const rootReducer = combineReducers({events, matches, user, teams});
